Add files via upload
A python script that can test large tables of data and give a ranking of which distribution it fits the best for a customizable set of distributions. Used to test over 2000 real world datasets in gas, oil, water flowrate and temperature, pressure, CO2 mol%, H2S mol%, HCO3 ppm, Cl- ppm in Pipelines and well production tubing. This is for informing decisions when using monte carlo corrosion models with options for distribution.